requirements.txt
django
djangorestframework
django-extensions

# test/debug
django-debug-toolbar
factory-boy
faker
